Refactor bottom navigation bar into a standalone component
- Extracted MainBottomBar and RecordIcon to a new file MainBottomBar.kt
- Updated MainTab model to support badge counts and notification dots
- Integrated Material 3 BadgedBox into the bottom bar
- Cleaned up AppNavigation.kt and verified build and tests
- Added helper method fromRoute to MainTab companion object
