+++ DouyinLite/app/src/main/java/com/app/douyin/pro/navigation/MainTab.kt
@@ -9,7 +9,9 @@ import androidx.compose.ui.graphics.vector.ImageVector
 sealed class MainTab(
     val route: String,
     val title: String,
-    val icon: ImageVector?
+    val icon: ImageVector?,
+    val badgeCount: Int = 0,
+    val showDot: Boolean = false
 ) {
     data object Home : MainTab(NavRoutes.HOME, "首页", Icons.Filled.Home)
     data object Friends : MainTab(NavRoutes.FRIENDS, "朋友", Icons.Filled.Person)
@@ -19,5 +21,9 @@ sealed class MainTab(
 
     companion object {
         val items = listOf(Home, Friends, Record, Inbox, Me)
+
+        fun fromRoute(route: String?): MainTab? {
+            return items.find { it.route == route }
+        }
     }
 }
